requirements.txt
pytest==8.0.1
selenium==4.17.2
allure-pytest=2.13.2
